+++ crates/ryft-core/src/types_v0/mod.rs
@@ -1,4 +1,1 @@
-pub mod array_type;
 pub mod r#type;
-
-pub use array_type::ArrayType;