+++ crates/ryft-core/src/types_v0/type.rs
@@ -5,10 +5,8 @@ use crate::{
     programs::{Constant, ConstantExpression, Program, ProgramType},
     tracing_v0::{Tracer, VariableTracer},
     types::data_type::DataType,
-    types::{Type, Typed},
+    types::{ArrayType, Type, Typed},
 };
-
-use super::array_type::ArrayType;
 
 // Our tracing module defines its own [DataType] and [Shape] types that ought to be decoupled from the corresponding
 // types that may be used by different kinds of [Parameter]s. This enables us to support multiple different backends