+++ crates/ryft/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
-- Updated the `ryft-mlir`, `ryft-pjrt`, and `ryft-xla-sys` dependencies.
+- Updated the `ryft-core`, `ryft-mlir`, `ryft-pjrt`, and `ryft-xla-sys` dependencies.
 
 ## [0.0.2] - 2026-03-02